Issue #1856: Fixed memory leak
