+++ src/app/modules/scalper-order-book/models/scalper-order-book-data-context.model.ts
@@ -39,6 +39,5 @@ export interface ScalperOrderBookDataContext {
 
   addLocalOrder(order: LocalOrder): void;
   removeLocalOrder(orderId: string): void;
-
   destroy(): void;
 }
